+++ src/engines/gs_engine_class.F90
@@ -312,6 +312,9 @@ contains
       use quasi_newton_updater_class,   only: quasi_newton_updater
       use newton_raphson_updater_class, only: newton_raphson_updater
 !
+      use citation_class,           only : citation
+      use citation_printer_class,   only : eT_citations
+!
       implicit none 
 !
       class(gs_engine), intent(in) :: engine 
@@ -323,6 +326,8 @@ contains
       class(amplitude_updater), allocatable :: t_updater
 !
       class(abstract_jacobian_transformer), allocatable :: transformer
+!
+      class(citation), allocatable :: reference
 !
       real(dp) :: relative_threshold
 !
@@ -365,6 +370,22 @@ contains
          if (input%is_keyword_present('multimodel newton', 'solver cc gs')) then 
 !
             transformer = approximate_jacobian_transformer('right')
+!
+            reference = citation(implementation = 'Multimodel Newton algorithm',                  &
+                                 journal        = 'J. Chem. Phys.',                               &
+                                 title_         = 'Accelerated multimodel Newton-type algorithms &
+                                                   &for faster convergence of ground and excited &
+                                                   &state coupled cluster equations',             &
+                                 volume         = '153',                                          &
+                                 issue          = '1',                                            &
+                                 pages          = '014104',                                       &
+                                 year           = '2020',                                         &
+                                 doi            = '10.1063/5.0010989',                            &
+                                 authors        = [character(len=25) :: 'Eirik F. Kjønstad',      &
+                                                                        'Sarai D. Folkestad',     &
+                                                                        'Henrik Koch'])
+!
+            call eT_citations%add(reference)
 !
          else 
 !